+++ Cereal/Cereal/src/Header.h
@@ -56,7 +56,7 @@ namespace Cereal {
 			if (!buffer.hasSpace(this->getSize())) return false;
 
 			buffer.writeBytes<unsigned short>(MAGIC_NUMBER);
-			buffer.writeBytes<byte>(databases.size());
+			buffer.writeBytes<byte>((byte)databases.size());
 
 			unsigned int offset = sizeof(short) + sizeof(byte) + (sizeof(unsigned int) * databases.size());
